contract_sale_generation_client_order_ref/test-requirements.txt
odoo-addon-contract_sale @ git+https://github.com/OCA/contract.git@refs/pull/915/head#subdirectory=setup/contract_sale
odoo-addon-contract_sale_generation @ git+https://github.com/OCA/contract.git@refs/pull/914/head#subdirectory=setup/contract_sale_generation
